put messages in package; remote redundant
