+++ src/fowl/_proto.py
@@ -27,6 +27,7 @@ from wormhole.cli.public_relay import RENDEZVOUS_RELAY as PUBLIC_MAILBOX_URL
 import wormhole.errors as wormhole_errors
 
 from .observer import When
+from .messages import *
 
 
 
@@ -51,135 +52,6 @@ def _sequential_id():
 
 
 allocate_connection_id = partial(next, _sequential_id())
-
-
-class FowlOutputMessage:
-    """
-    An information message from fowld to the controller
-    """
-
-
-class FowlCommandMessage:
-    """
-    A command from the controller to fowld
-    """
-
-
-# if we had ADT / Union types, these would both be that -- is this as
-# close as we can get in Python?
-
-
-@frozen
-class Welcome(FowlOutputMessage):
-    """
-    We have connected to the Mailbox Server and received the
-    Welcome message.
-    """
-    # open-ended information from the server
-    welcome: dict
-
-
-@frozen
-class WormholeClosed(FowlOutputMessage):
-    """
-    The wormhole has been terminated.
-    """
-    result: str
-
-
-@frozen
-class AllocateCode(FowlCommandMessage):
-    """
-    Create a fresh code on the server
-    """
-    length: Optional[int] = None
-
-
-@frozen
-class SetCode(FowlCommandMessage):
-    """
-    Give a code we know to the server
-    """
-    code: str
-
-
-@frozen
-class CodeAllocated(FowlOutputMessage):
-    """
-    The secret wormhole code has been determined
-    """
-    code: str
-
-
-@frozen
-class PeerConnected(FowlOutputMessage):
-    """
-    We have evidence that the peer has connected
-    """
-    # hex-encoded 32-byte hash output (should match other peer)
-    verifier: str
-    versions: dict
-
-
-@frozen
-class LocalListener(FowlCommandMessage):
-    """
-    We wish to open a local listener.
-    """
-    listen: str  # Twisted server-type endpoint string
-    connect: str  # Twisted client-type endpoint string
-
-
-@frozen
-class RemoteListener(FowlCommandMessage):
-    """
-    We wish to open a listener on the peer.
-    """
-    listen: str  # Twisted server-type endpoint string
-    connect: str  # Twisted client-type endpoint string
-
-
-@frozen
-class Listening(FowlOutputMessage):
-    """
-    We have opened a local listener.
-
-    Any connections to this listener will result in a subchannel and a
-    connect on the other side (to "connected_endpoint"). This message
-    may result from a LocalListener or a RemoteListener command. This
-    message will always appear on the side that's actually listening.
-    """
-    listen: str  # Twisted server-type endpoint string
-    connect: str  # Twisted client-type endpoint string
-
-
-@frozen
-class LocalConnection(FowlOutputMessage):
-    """
-    Something has connected to one of our listeners
-    """
-    id: int
-
-
-@frozen
-class IncomingConnection(FowlOutputMessage):
-    """
-    The other side is requesting we open a connection
-    """
-    id: int
-    endpoint: str
-
-
-@frozen
-class BytesIn(FowlOutputMessage):
-    id: int
-    bytes: int
-
-
-@frozen
-class BytesOut(FowlOutputMessage):
-    id: int
-    bytes: int
 
 
 @frozen
@@ -1262,115 +1134,6 @@ class FowlWormhole:
             raise KeyError(
                 "Unknown command '{}'".format(cmd["kind"])
             )
-
-
-##XXXX
-# okay so if we really want "a sans-io core" then all
-# the "messages" have to be processed synchronously (by the state-machine)
-# so what _ARE_ the in / out messages?
-# we have two categories:
-# - "stuff for the wormhole" (async calls to / from .. _DelegatedWormhole instance)
-# - "stuff for the controller" (stdin/stdout)
-#
-# I _think_ it's okay to literally "be" the "delegate" (since all sync.)
-# so, brainstorming through this "in code":
-
-### wormhole messages
-# "to" the wormhole messages:
-
-class ToWormhole:
-    pass
-
-
-# is this maybe just "outside" the whole thing?
-# "somehow" we cause code_allocated() to be called (either "given to
-# us" or "we asked workhole for it")
-
-@frozen
-class SetCode(ToWormhole):
-    code: str
-
-
-@frozen
-class AllocateCode(ToWormhole):
-    code_length: int = 2
-
-
-@frozen
-class SendMessage(ToWormhole):
-    plaintext: str
-
-
-@frozen
-class CloseWormhole(ToWormhole):
-    pass
-
-
-# "from" the wormhole messages (eveything the "delegate" calls, wormhole_*)
-# (do we "bother" with these, or just hook the "delegate" directly in..?)
-
-# hrrrmmm .. from the 'state machine' perspective, we just use @m.input for all our input I guess?
-# otherwise, we just have a "process_input_message" that will do dumb type introspection
-# and call an actual @m.input method
-# ...so it's only for "emit_*" things that we have "actual message classes"
-# ..._OR_ we could I guess do a "delegate" sort of thing, like magic-wormhole
-
-
-### sans-io story here is this:
-### stuff "coming in" are @m.input() methods
-###   - (things on standard-in start as FowlInputMessages via ".command()")
-### stuff "going out" is a FowlOutputMessage subclass
-
-
-class FowlOutputMessage:
-    pass
-
-
-@frozen
-class GotMessageFromPeer(FowlOutputMessage):
-    plaintext: str
-
-
-@frozen
-class SendMessageToPeer(FowlOutputMessage):
-    plaintext: str
-
-
-@frozen
-class AllocateCode(FowlCommandMessage):
-    """
-    Create a fresh code on the server
-    """
-    length: Optional[int] = None
-
-
-@frozen
-class SetCode(FowlCommandMessage):
-    """
-    Give a code we know to the server
-    """
-    code: str
-
-
-@frozen
-class LocalListener(FowlCommandMessage):
-    """
-    We wish to open a local listener.
-    """
-    listen: str  # Twisted server-type endpoint string
-    connect: str  # Twisted client-type endpoint string
-
-
-@frozen
-class RemoteListener(FowlCommandMessage):
-    """
-    We wish to open a listener on the peer.
-    """
-    listen: str  # Twisted server-type endpoint string
-    connect: str  # Twisted client-type endpoint string
-
-
-# -to outside (i.e. "emit on stdout")
 
 
 # FowlDaemon is the state-machine
